everything except for tooltip styling and room data perpetuation
