+++ front/src/types/room/shared.ts
